Завершена настройка админ-панели. Поправлены модели форм: в базовую модель добавлено поле, обработан ли запрос.

+++ watches/main/models.py
@@ -84,7 +84,7 @@ class Watch(Base):
             'Обязательное поле.'
         )
     )
-    price = models.PositiveSmallIntegerField(
+    price = models.PositiveIntegerField(
         validators=[
             MinValueValidator(
                 MIN_WATCH_PRICE,
